requirements.txt
langchain>=0.1.0
langchain-core>=0.1.0
langchain-groq>=0.1.0
python-dotenv>=1.0.0
faiss-cpu>=1.7.4
pydantic>=2.0.0
python-decouple>=3.8
typing-extensions>=4.5.0